perf(etfs): split price-chart slice into its own /chart-data endpoint
The main ETF detail page fetched everything through one /full-render call,
so the report body (analysis, holdings, competition) had to wait on
ensureEtfPriceHistoryIsFresh, which can hit Yahoo over the network.

Move the price-chart + performance-metrics slice into a new /chart-data
endpoint and give the chart-tabs Suspense its own promise, so the report
body streams as soon as the DB read finishes and the chart streams
independently. The remaining slices (about, radar, below-fold) keep sharing
/full-render since they all read from one ETF-row + analysis query.

- New: etfs-v1/.../[etf]/chart-data route ({ priceHistory, performanceMetrics })
- Trim /full-render: drop priceHistory/performanceMetrics + the Yahoo refresh
  and category-peer query (and now-unused imports/helper)
- Page: fetchEtfChartData wired to the chart-tabs Suspense; same etfAndExchangeTag
- Cache: add /chart-data to CloudFront cached paths + the per-ETF edge purge
