+++ insights-ui/src/app/api/[spaceId]/etfs-v1/exchange/[exchange]/[etf]/full-render/route.ts
@@ -11,22 +11,14 @@
  */
 import { EtfAnalysisResponse, EtfCategoryAnalysisResultResponse } from '@/app/api/[spaceId]/etfs-v1/exchange/[exchange]/[etf]/analysis/route';
 import { EtfPortfolioHoldingsResponse } from '@/app/api/[spaceId]/etfs-v1/exchange/[exchange]/[etf]/portfolio-holdings/route';
-import { PriceHistoryResponse } from '@/app/api/[spaceId]/tickers-v1/exchange/[exchange]/[ticker]/price-history/route';
 import { EtfFastResponse } from '@/app/api/[spaceId]/etfs-v1/exchange/[exchange]/[etf]/route';
 import type { EtfFinancialInfoResponse, EtfScoresResponse, SimilarEtf } from '@/types/etf/etf-detail-response-types';
 import { getEtfWhereClause, serializeBigIntFields } from '@/app/api/[spaceId]/etfs-v1/etfApiUtils';
 import { prisma } from '@/prisma';
 import type { EtfApplicableInvestorGoals, EtfCompetitionResponse, EtfCompetitor, EtfKeyFactsFlagAssessment } from '@/types/etf/etf-analysis-types';
 import { KoalaGainsSpaceId } from '@/types/koalaGainsConstants';
-import { PriceHistoryPoint } from '@/types/prismaTypes';
 import { CompetitionAnalysis } from '@/types/public-equity/analysis-factors-types';
-import { ensureEtfPriceHistoryIsFresh } from '@/utils/etf-price-history-utils';
-import {
-  buildEtfPerformanceMetricsPayload,
-  parsePercentString,
-  type EtfPerformanceMetricFields,
-  type EtfPerformanceMetricsPayload,
-} from '@/utils/etf-performance-metrics-utils';
+import { parsePercentString } from '@/utils/etf-performance-metrics-utils';
 import { withErrorHandlingV2 } from '@dodao/web-core/api/helpers/middlewares/withErrorHandling';
 import { NextRequest } from 'next/server';
 
@@ -66,8 +58,6 @@ export interface EtfFullRenderResponse {
   similarEtfs: SimilarEtf[];
   portfolioHoldings: EtfPortfolioHoldingsResponse;
   competition: EtfCompetitionResponse | null;
-  priceHistory: PriceHistoryResponse | null;
-  performanceMetrics: EtfPerformanceMetricsPayload | null;
   keyFacts: EtfKeyFactsReportResponse | null;
   keyMetrics: EtfKeyMetricsResponse;
 }
@@ -93,8 +83,6 @@ const EMPTY: EtfFullRenderResponse = {
   similarEtfs: [],
   portfolioHoldings: { holdings: null, updatedAt: null },
   competition: null,
-  priceHistory: null,
-  performanceMetrics: null,
   keyFacts: null,
   keyMetrics: EMPTY_KEY_METRICS,
 };
@@ -128,9 +116,8 @@ async function getHandler(
   if (!etfRecord) return EMPTY;
 
   const etfId = etfRecord.id;
-  const focalCategory = etfRecord.stockAnalyzerInfo?.category ?? null;
 
-  const [analysisRows, similarHydrated, competitorHydrated, priceInfo, categoryPeers] = await Promise.all([
+  const [analysisRows, similarHydrated, competitorHydrated] = await Promise.all([
     prisma.etfCategoryAnalysisResult.findMany({
       where: { etfId, spaceId: where.spaceId },
       include: {
@@ -141,8 +128,6 @@ async function getHandler(
     }),
     hydrateSimilarEtfs(etfRecord.spaceId, etfRecord.similarEtfs),
     hydrateCompetitors(etfRecord.vsCompetition?.competitionAnalysisArray as unknown as CompetitionAnalysis[] | undefined),
-    ensureEtfPriceHistoryIsFresh(etfRecord).catch(() => null),
-    fetchCategoryPeerAnalyzerInfo(etfRecord.spaceId, focalCategory, etfId),
   ]);
 
   // Shape responses to match the original per-route payloads so the main page
@@ -251,22 +236,6 @@ async function getHandler(
       }
     : null;
 
-  const performanceMetrics: EtfPerformanceMetricsPayload | null = etfRecord.stockAnalyzerInfo
-    ? buildEtfPerformanceMetricsPayload(etfRecord.stockAnalyzerInfo, categoryPeers, focalCategory)
-    : null;
-
-  const priceHistory: PriceHistoryResponse | null =
-    priceInfo &&
-    (((priceInfo.dailyData as unknown as PriceHistoryPoint[] | null)?.length ?? 0) > 0 ||
-      ((priceInfo.weeklyData as unknown as PriceHistoryPoint[] | null)?.length ?? 0) > 0)
-      ? {
-          symbol: etfRecord.symbol,
-          currency: priceInfo.currency ?? null,
-          daily: (priceInfo.dailyData as unknown as PriceHistoryPoint[] | null) ?? [],
-          weekly: (priceInfo.weeklyData as unknown as PriceHistoryPoint[] | null) ?? [],
-        }
-      : null;
-
   return {
     etf: etfRaw,
     financialInfo,
@@ -275,8 +244,6 @@ async function getHandler(
     similarEtfs: similarHydrated,
     portfolioHoldings,
     competition,
-    priceHistory,
-    performanceMetrics,
     keyFacts,
     keyMetrics,
   };
@@ -307,42 +274,6 @@ function extractMaxDrawdown(riskPeriods: unknown): number | null {
     if (parsed !== null) return parsed;
   }
   return null;
-}
-
-/**
- * Pull only the cagr/return fields for the focal ETF's category peers (same
- * `spaceId`, same `category`, excluding the focal ETF itself). Returns
- * `EtfStockAnalyzerInfo`-shaped partials whose unread fields are filled with
- * `null` so the shared averaging util can read them safely.
- *
- * Returns `[]` if the focal ETF has no category — the chart will then render
- * with the ETF-only series.
- */
-async function fetchCategoryPeerAnalyzerInfo(spaceId: string, category: string | null, focalEtfId: string): Promise<EtfPerformanceMetricFields[]> {
-  if (!category) return [];
-
-  return prisma.etfStockAnalyzerInfo.findMany({
-    where: {
-      etf: { spaceId, NOT: { id: focalEtfId } },
-      category,
-    },
-    select: {
-      cagr1y: true,
-      cagr3y: true,
-      cagr5y: true,
-      cagr10y: true,
-      cagr15y: true,
-      cagr20y: true,
-      return1m: true,
-      return6m: true,
-      return1y: true,
-      return3y: true,
-      return5y: true,
-      return10y: true,
-      return15y: true,
-      return20y: true,
-    },
-  });
 }
 
 async function hydrateSimilarEtfs(spaceId: string, stored: Array<{ id: string; symbol: string; exchange: string; sortOrder: number }>): Promise<SimilarEtf[]> {
